vulnerable-apps/web-app/requirements.txt
flask>=2.0.0
flask-mysql>=1.5.0
mysql-connector-python>=8.0.0
